```
SQA's baked queries prepares the queries in advance, caching yada yada.
Makes thing a bit faster.
Also bigger speedup included is a shoddy cache for the total torrent
count (only applied to baked queries currently). Caching the value for a
few dozen seconds shaves off some wasted time, as it's mostly just used
for pagination.
```

Yeah yeah, get baked, ha ha.

Simple benchmarks
```
Before baked
Page   1: min: 59.0ms max: 72.9ms avg: 67.1ms
Page   2: min: 56.2ms max: 97.6ms avg: 66.1ms
Page   3: min: 60.1ms max: 92.4ms avg: 69.3ms
Page   4: min: 59.9ms max: 88.0ms avg: 68.8ms
Page   5: min: 60.1ms max: 90.9ms avg: 67.1ms

With baked (no count cache, COUNT_CACHE_DURATION=0)
Page   1: min: 56.6ms max: 69.2ms avg: 61.8ms
Page   2: min: 51.0ms max: 75.5ms avg: 62.1ms
Page   3: min: 52.8ms max: 76.5ms avg: 59.9ms
Page   4: min: 49.4ms max: 75.8ms avg: 59.9ms
Page   5: min: 53.9ms max: 89.0ms avg: 63.3ms

With baked and count cache
Page   1: min: 37.8ms max: 52.0ms avg: 43.9ms
Page   2: min: 38.8ms max: 65.7ms avg: 45.4ms
Page   3: min: 39.4ms max: 72.3ms avg: 45.7ms
Page   4: min: 40.7ms max: 59.5ms avg: 46.0ms
Page   5: min: 41.6ms max: 77.4ms avg: 49.1ms
```

Maybe also look at #594